requirements.txt
pandas
sklearn
numpy
django
djangorestframework
